Included the PackageDev fixes to User/XML.sublime-settings due
The XML package includes a `XML.sublime-settings` file that sets
`hidden_extensions` to some of the extension we want to highlight
with our package. There is currently no other way to override
this, so we manually remove these extensions from the setting
with a plugin.
MODIFIED_VALUE = ["rss", "vcproj", "tmLanguage", "tmTheme", "tmSnippet", "dae"]
...
The XML package includes a `XML.sublime-settings` file that sets
`hidden_extensions` to some of the extension we want to highlight
with our package. There is currently no other way to override
this, so we manually remove these extensions from the setting
with a plugin.
MODIFIED_VALUE = ["rss", "vcproj", "tmLanguage", "tmTheme", "tmSnippet", "dae"]

https://github.com/SublimeText/PackageDev/blob/89cc1146b48604c0e0a38764c86397e0723ed614/xml_hidden_extensions_hotfix.py